[theme] move .*Font tests to font_test.go
